==> code/14_process_future_predictions.R <==
#----------------------------------#
#change bloom date
#----------------------------------#

library(chillR)
library(LarsChill)
library(tidyverse)


#get the name of the cultivars considered in this study

adamedor <-read.csv('data/combined_phenological_data_adamedor_clean.csv') 

f50_cult <- adamedor %>% 
  dplyr::filter(species %in% c(c('Sweet Cherry', 'Pistachio', 'Pear', 'Apricot', 'Almond'))) %>% 
  dplyr::select(species, cultivar, location, year, flowering_f50) %>% 
  stats::na.omit() %>% 
  group_by(species, cultivar) %>% 
  summarise(n = n()) %>% 
  dplyr::filter(n >= 20) %>% 
  dplyr::pull(cultivar)


f5_cult <- adamedor %>% 
  dplyr::filter(species %in% c(c('Apple', 'European plum', 'Japanese plum'))) %>% 
  dplyr::select(species, cultivar, location, year, begin_flowering_f5) %>% 
  stats::na.omit() %>% 
  group_by(species, cultivar) %>% 
  summarise(n = n()) %>% 
  dplyr::filter(n >= 20 )%>% 
  dplyr::pull(cultivar)

cultivar_n <- c(f50_cult, f5_cult)



#read the predictions for 2015
pheno_2015 <- read.csv('data/projected_bloomdates_ensemble_historic_scenarios.csv') %>% 
  mutate(species = recode(species, `Japanese Plum` = 'European Plum')) %>% 
  filter(cultivar %in% cultivar_n)

#read the validation performance
performance <- read.csv('data/performance_fitted_models.csv') %>% 
  filter(cultivar %in% cultivar_n, 
         split == 'Validation') %>% 
  mutate(species = recode(species, `Japanese Plum` = 'European Plum'))

performance %>% 
  group_by(species) %>% 
  summarize(med = median(rmse),
            quan_05 = quantile(rmse, probs = 0.05),
            quan_95 = quantile(rmse, probs = 0.95))

performance %>% 
  group_by(species) %>% 
  summarize(med = median(rpiq_adj),
            quan_05 = quantile(rpiq_adj, probs = 0.05),
            quan_95 = quantile(rpiq_adj, probs = 0.95))





# #read the different batches of the future bloom dates
# f <- list.files('data/predictions_future_v2/', full.names = TRUE) %>% 
#   purrr::map(read.csv, .progress = TRUE) %>% 
#   do.call(rbind, .) 
# 
# 
# 
# prepare_conf <- function(m1, conf,  NA_threshold = 5){
#   apply(m1, MARGIN = 1, FUN = function(x){
#     if(sum(is.na(x)) >= NA_threshold){
#       return(rep(NA, length(x)))
#     } 
#     
#     if(any(is.na(x))){
#       conf_adjust <- conf
#       conf_adjust[is.na(x)] <- 0
#       
#       #now give the rescaled values of confidence between 0 and 1
#       conf_adjust <- conf_adjust / sum(conf_adjust)
#       
#       return(conf_adjust)
#       
#     }
#     
#     return(conf / sum(conf))
#   })
# }
# prepare_m1 <- function(m1){
#   apply(m1, MARGIN = 1, FUN = function(x){
#     x[is.na(x)] <- 0
#     return(x)
#   }) %>% 
#     t() %>% 
#     return()
# }
# 
# 
# 
# 
# #split species_cultivar column
# f$cultivar <- str_split(f$spec_cult, '_') %>% 
#   purrr::map_chr(2)
# 
# #filter for cultivars which were included in the study
# f <- f %>% 
#   filter(cultivar %in% cultivar_n)
# 
# 
# #get the cultivars to iterate over
# cases <- unique(f$spec_cult) 
# f <- data.table::data.table(f)
# 
# #calculate the weighted prediction
# weighted_pred <- purrr::map(cases, function(case){
#   
#   #extract the weights
#   conf <- performance %>% 
#     mutate(spec_cul = paste0(species,'_', cultivar)) %>% 
#     filter(spec_cul == case) %>% 
#     arrange(repetition) %>% 
#       pull(rpiq_adj)
#   
#   #arrange as a matrix, adjust the weights so that they add up to 1
#   conf_matrix <- f %>% 
#       filter(spec_cult == case) %>% 
#       select(X1:X10) %>% 
#       as.matrix() %>% 
#       prepare_conf(conf = conf)
#   
#   #arrange the predictions as a matrix, remove all predictions if more than 5 NAs
#   m1 <-f %>% 
#       filter(spec_cult == case) %>% 
#       select(X1:X10) %>% 
#       as.matrix() %>% 
#       prepare_m1()
# 
#   #calculate the weighted sum
#   test <- purrr::map_dbl(1:ncol(conf_matrix), function(i) m1[i,] %*% conf_matrix[,i])
#     
#     return(test)
#   }, .progress = TRUE) %>% 
#   do.call(c,.) 
# 
# 
# pred_info <- purrr::map(cases, function(case){
#   
#   f %>% 
#     filter(spec_cult == case) %>% 
#     select(spec_cult, location, gcm, ssp, scenario_year, year_id) %>% 
#     return()
# }, .progress = TRUE) %>% 
#   do.call(rbind,.) 
# 
#   #calculate the sd of the prediction
# f_test <- f %>% 
#   filter(spec_cult %in% cases) %>% 
#   select(X1:X10) %>% 
#   apply(MARGIN = 1, function(r) sd(r, na.rm = TRUE)) %>% 
#   unlist() %>% 
#   cbind(f)
# 
# 
# #add the weighted prediction to the pred_info and then merge it with the f_test object to have everything together
# f_out <- pred_info %>% 
#   mutate(weighted_pred = weighted_pred) %>% 
#   merge(f_test, by = c('spec_cult', 'location', 'gcm', 'ssp', 'scenario_year', 'year_id'))
# 
# write.csv(f_out, file = 'data/projected_bloomdates_future.csv', row.names = FALSE)
# 
f_out <- read.csv('data/projected_bloomdates_future.csv')

#rm(f, f_test, sd, weighted_pred)

#memory.limit(size = 10000)
f_out <- data.table::as.data.table(f_out)
gc()



#split the column into species and cultivar
f_out$species <- sapply(strsplit(as.character(f_out$spec_cult), "_"), function(x) x[1])
f_out$cultivar <- sapply(strsplit(as.character(f_out$spec_cult), "_"), function(x) x[2])

#subset so that we only have cultivars we included in the study
f_out <- f_out[cultivar %in% cultivar_n]

#make sure Japanese Plum is labelled as European plum
f_out$species <- gsub("Japanese Plum", 'European Plum', f_out$species)


names(f_out)[names(f_out) == "."] <- "sd"
names(f_out)[names(f_out) == "weighted_pred"] <- "pheno_predicted"


#make plot to compare projections of local fitted models and "foreign" models
master_fitting_data <- read.csv('data/master_phenology_repeated_splits.csv') %>%
  filter(repetition == 1) %>%
  mutate(species = gsub('Japanese Plum', 'European Plum', species)) %>%
  group_by(species, cultivar) %>%
  summarize(loc = unique(location)) %>%
  mutate(spec_cult_loc = paste(species, cultivar, loc, sep = '_'),
         species_location = paste(species, loc, sep = '_')) %>%
  filter(cultivar %in% cultivar_n)

master_fitting_data %>%
  filter(species == 'Almond',
         cultivar %in% cultivar_n)%>%
  group_by(cultivar) %>%
  summarize(n = length(unique(loc))) %>%
  filter(n == 3)


f_out$spec_cult_loc <- paste(f_out$species, f_out$cultivar, f_out$location, sep = '_')
f_out$local_cultivar <- ifelse(f_out$spec_cult_loc %in% master_fitting_data$spec_cult_loc, yes = TRUE, no = FALSE)

f_out$fail <- is.na(f_out$pheno_predicted)
fail_intermediate <- f_out[, .(fail_rate = round((sum(fail) / length(fail))*100, digits = 2)),
                           by = .(species, location, ssp, gcm, scenario_year, local_cultivar)]
fail_summarized <- fail_intermediate[,.(median_fail = median(fail_rate)),
                                     by = .(species, location, ssp, scenario_year, local_cultivar)]

#summarize the fail also on a cultivar level
fail_cult <- f_out[, .(fail_rate = round((sum(fail) / length(fail))*100, digits = 2)),
                   by = .(species, cultivar, location, ssp, gcm, scenario_year, local_cultivar)]



median_future <- f_out[, .(med_future = median(pheno_predicted, na.rm = TRUE),
                 sd_future = sd(pheno_predicted, na.rm = TRUE)),
               by = .(species, location, ssp, gcm, scenario_year)]


#this is for the shift plot
median_future_local_vs_foreign <- f_out[, .(med_future = median(pheno_predicted, na.rm = TRUE),
                           sd_future = sd(pheno_predicted, na.rm = TRUE)),
                       by = .(species, location, ssp, gcm, scenario_year, local_cultivar)]


#this is needed for the point +- sd plot
median_future_local_vs_foreign_allgcm <- f_out[, .(med_future = median(pheno_predicted, na.rm = TRUE),
                                            sd_future = sd(pheno_predicted, na.rm = TRUE)),
                                        by = .(species, location, ssp, scenario_year, local_cultivar)]

rm(f_out)
gc()

fail_cult %>%
  filter(species == 'Apricot', location == 'Cieza') %>%
  ggplot(aes(x = cultivar, y = fail_rate, fill = ssp)) +
  geom_boxplot() +
  facet_grid(local_cultivar~scenario_year)

t <- fail_cult %>%
  filter(species == 'Apricot', location == 'Cieza') %>%
  group_by(cultivar, local_cultivar, ssp, scenario_year) %>%
  summarise(med_fail = median(fail_rate)) %>%
  filter(local_cultivar)

fail_cult %>%
  filter(species == 'Pistachio', location == 'Sfax') %>%
  ggplot(aes(x = cultivar, y = fail_rate, fill = ssp)) +
  geom_boxplot() +
  facet_grid(local_cultivar~scenario_year)

loc_almond_cult <- c('Abiodh Ras Djebel', 'Faggoussi', 'Garnghzel',
              'Tozeur 4', 'Fakhfekh', 'Khoukhi', 'Tozeur 1',
              'Achak', 'Ksontini', 'Zahaf', 'Abiodh de Sfax',
              'Tozeur 2', 'Achaak')
for_almond_cult <- c('Tarragona', 'Cristomorto', 'Nonpareil',
                     'Pizzuta', 'Genco Taronto', 'Fournat de Breznaud',
                     'Soukaret', 'Trell', 'Bonifacio', 'Mazzetto',
                     'Ferraduel', 'Ferragnes', 'Ne Plus Ultra', 'Avola',
                     'Tuono Taronto', 'Rechelle Taronto', 'Doree',
                     'Bruantine', 'Cavaliera', 'Fasciuneddu',
                     'Ramlet R249', 'Montrone', 'Languedoc',
                     'Malaguena', 'Ramlet R250', 'Dorée')

fail_cult %>%
  filter(species == 'Almond', location == 'Sfax') %>%
  ggplot(aes(x = cultivar, y = fail_rate, fill = ssp)) +
  geom_boxplot() +
  facet_grid(local_cultivar~scenario_year)

fail_cult %>%
  filter(species == 'Almond', location == 'Sfax') %>%
  mutate(loc_almond = ifelse(cultivar %in% loc_almond_cult,
                             yes = TRUE,
                             no = ifelse(cultivar %in% for_almond_cult, yes = FALSE, no = NA))) %>%
  filter(scenario_year == 2050, local_cultivar == TRUE) %>%
  ggplot(aes(x = cultivar, y = fail_rate, fill = ssp)) +
  geom_boxplot() +
  facet_grid(ssp~loc_almond, scales = 'free_x')

fail_cult %>%
  filter(species == 'Almond', location == 'Sfax') %>%
  mutate(loc_almond = ifelse(cultivar %in% loc_almond_cult,
                             yes = TRUE,
                             no = ifelse(cultivar %in% for_almond_cult, yes = FALSE, no = NA))) %>%
  filter(scenario_year == 2085, local_cultivar == TRUE) %>%
  ggplot(aes(x = cultivar, y = fail_rate, fill = ssp)) +
  geom_boxplot() +
  facet_grid(ssp~loc_almond, scales = 'free_x')






t <- pheno_2015 %>%
  mutate(fail = is.na(pheno_predicted)) %>%
  group_by(species, cultivar, location) %>%
  summarize(fail_rate = (sum(fail) / n())*100) %>%
  filter(location == 'Zaragoza', species == 'Sweet Cherry')

pheno_2015 %>%
  group_by(species, cultivar, location) %>%
  mutate(fail_rate = (sum(is.na(pheno_predicted)) / n())*100,
         ssp = '2015',
         scenario_year = 2015) %>%
  rbind(fail_cult) %>%
  filter(location == 'Meknes', species == 'Apple') %>%
  ggplot(aes(x = cultivar, y = fail_rate, fill = ssp)) +
  geom_boxplot() +
  facet_grid(~scenario_year)

pheno_2015 %>%
  group_by(species, cultivar, location) %>%
  mutate(fail_rate = (sum(is.na(pheno_predicted)) / n())*100,
         ssp = '2015',
         scenario_year = 2015) %>%
  rbind(fail_cult) %>%
  filter(location == 'Cieza', species == 'Apricot') %>%
  ggplot(aes(x = cultivar, y = fail_rate, fill = ssp)) +
  geom_boxplot() +
  facet_grid(~scenario_year)

pheno_2015 %>%
  group_by(species, cultivar, location) %>%
  mutate(fail_rate = (sum(is.na(pheno_predicted)) / n())*100,
         ssp = '2015',
         scenario_year = 2015) %>%
  rbind(fail_cult) %>%
  filter(location == 'Zaragoza', species == 'Sweet Cherry') %>%
  ggplot(aes(x = cultivar, y = fail_rate, fill = ssp)) +
  geom_boxplot() +
  facet_grid(~scenario_year)





#calculate median for both
median_2015 <- pheno_2015 %>%
  group_by(species, location) %>%
  summarise(med_current = median(pheno_predicted, na.rm = TRUE),
            sd_current = sd(pheno_predicted, na.rm = TRUE))

median_2015_locfor <- pheno_2015 %>%
  mutate(spec_cult_loc = paste(species, cultivar, location, sep = '_'),
         local_cultivar = ifelse(spec_cult_loc %in% master_fitting_data$spec_cult_loc, yes = TRUE, no = FALSE)) %>%
  group_by(species, location, local_cultivar) %>%
  summarise(med_current = median(pheno_predicted, na.rm = TRUE),
            sd_current = sd(pheno_predicted, na.rm = TRUE))


shift_df <- merge(median_2015, median_future, by = c('species','location')) %>%
  mutate(shift_bloom = round(med_future - med_current, digits = 2))

shift_table <- shift_df %>%
  group_by(species, location, scenario_year, ssp) %>%
  summarise(min_shift = min(shift_bloom, na.rm = TRUE),
            max_shift = max(shift_bloom, na.rm = TRUE))
# 
# write.csv(shift_table, 'data/summary_shift_bloom_date_text_v2.csv', row.names = FALSE)
# write.csv(shift_df, 'data/summary_shift_bloom_date_text_v2_inidvidual-gcm.csv', row.names = FALSE)

shift_table <- read.csv('data/summary_shift_bloom_date_text_v2.csv')
shift_df <- read.csv('data/summary_shift_bloom_date_text_v2_inidvidual-gcm.csv')



#merge the two
shift_df %>% 
  mutate(dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         spec = as.numeric(factor(species, levels = c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))),
         loc = as.numeric(factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax'))),
         med_future = ifelse(is.na(med_current), yes = NA, no = med_future)) %>% 
  filter(scenario_year == '2050') %>% 
  ggplot(ggplot2::aes(y = loc)) +
  geom_rect(aes(xmin = med_current, xmax = med_future, ymin = loc + dodge_low, ymax = loc + dodge_up, fill = ssp)) +
  geom_point(aes(x = med_future, y = loc + ((dodge_low + dodge_up)/2)),
             col = 'black',
             show.legend = FALSE, shape = 21) + 
  geom_rect(aes(xmin = med_current - 2.5, xmax = med_current + 2.5, ymax = loc - 0.4, ymin = loc + 0.4, fill = 'Simulation 2015'),  size = 2) +
  # geom_bar(stat = 'identity', position = 'dodge') +
  facet_grid(~species, scales = 'free_y', space = 'free_y') +
  scale_color_manual(values = c("#56B4E9", "#009E73","#F0E442",  "#E69F00"))+
  scale_fill_manual(values = c('black', "#56B4E9", "#009E73","#F0E442",  "#E69F00", 'grey70'))+
  theme_bw(base_size = 15) +
  scale_x_continuous(breaks = c(1,  60, 121, 182), 
                     labels = c('Jan', 'Mar', 'May', 'Jul'),
                     minor_breaks = c(1, 32, 60, 91, 121, 152, 182)) +
  scale_y_reverse(breaks = 1:6, labels =c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax')) +
  coord_cartesian(xlim = c(1, 182)) +
  theme(axis.text.x = ggplot2::element_text(angle = 45, hjust = 1), legend.position = 'bottom') +
  ylab('') +
  xlab('Median Predicted Bloom Date') +
  guides(fill=ggplot2::guide_legend(title="Weather Scenario"))
ggplot2::ggsave('figures/paper/change_med_bloom_2050_v1_updated_gcms.jpeg', device = 'jpeg',
                height = 40, width = 31, units = 'cm')


shift_df %>% 
  mutate(dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         spec = as.numeric(factor(species, levels = c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))),
         loc = as.numeric(factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax'))),
         med_future = ifelse(is.na(med_current), yes = NA, no = med_future)) %>% 
  filter(scenario_year == '2085') %>% 
  ggplot(ggplot2::aes(y = loc)) +
  geom_rect(aes(xmin = med_current, xmax = med_future, ymin = loc + dodge_low, ymax = loc + dodge_up, fill = ssp)) +
  geom_point(aes(x = med_future, y = loc + ((dodge_low + dodge_up)/2)),
             col = 'black',
             show.legend = FALSE, shape = 21) + 
  geom_rect(aes(xmin = med_current - 2.5, xmax = med_current + 2.5, ymax = loc - 0.4, ymin = loc + 0.4, fill = 'Simulation 2015'),  size = 2) +
  # geom_bar(stat = 'identity', position = 'dodge') +
  facet_grid(~species, scales = 'free_y', space = 'free_y') +
  scale_color_manual(values = c("#56B4E9", "#009E73","#F0E442",  "#E69F00"))+
  scale_fill_manual(values = c('black', "#56B4E9", "#009E73","#F0E442",  "#E69F00", 'grey70'))+
  theme_bw(base_size = 15) +
  scale_x_continuous(breaks = c(1,  60, 121, 182), 
                     labels = c('Jan', 'Mar', 'May', 'Jul'),
                     minor_breaks = c(1, 32, 60, 91, 121, 152, 182)) +
  scale_y_reverse(breaks = 1:6, labels =c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax')) +
  coord_cartesian(xlim = c(1, 182)) +
  theme(axis.text.x = ggplot2::element_text(angle = 45, hjust = 1), legend.position = 'bottom') +
  ylab('') +
  xlab('Median Predicted Bloom Date') +
  guides(fill=ggplot2::guide_legend(title="Weather Scenario"))
ggplot2::ggsave('figures/paper/change_med_bloom_2085_v1_updated_gcms.jpeg', device = 'jpeg',
                height = 40, width = 31, units = 'cm')

shift_df %>% 
  mutate(dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         location = factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax')),
         spec = as.numeric(factor(species, levels = c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))),
         loc = as.numeric(factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax'))),
         med_future = ifelse(is.na(med_current), yes = NA, no = med_future)) %>% 
  filter(scenario_year == '2050') %>% 
  ggplot(ggplot2::aes(y = spec)) +
  geom_rect(aes(xmin = med_current, xmax = med_future, ymin = spec + dodge_low, ymax = spec + dodge_up, fill = ssp)) +
  geom_point(aes(x = med_future, y = spec + ((dodge_low + dodge_up)/2)),
             col = 'black',
             show.legend = FALSE, shape = 4, size = 1) + 
  geom_rect(aes(xmin = med_current - 0.5, xmax = med_current + 0.5, ymax = spec - 0.4, ymin = spec + 0.4, fill = 'Simulation 2015'),  size = 2) +
  # geom_bar(stat = 'identity', position = 'dodge') +
  facet_grid(location~., scales = 'free_y', space = 'free_y') +
  scale_color_manual(values = c("#56B4E9", "#009E73","#F0E442",  "#E69F00"))+
  scale_fill_manual(values = c('black', "#56B4E9", "#009E73","#F0E442",  "#E69F00", 'grey70'))+
  theme_bw(base_size = 15) +
  scale_x_continuous(breaks = c(32,  60, 91, 121, 152), 
                     labels = c('Feb', 'Mar', 'Apr', 'May', 'Jun'),
                     minor_breaks = c(1, 32, 60, 91, 121, 152)) +
  scale_y_reverse(breaks = rep(1:7,6), labels = rep(c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'), 6)) +
  coord_cartesian(xlim = c(32, 155)) +
  theme(axis.text.x = ggplot2::element_text(angle = 45, hjust = 1), legend.position = 'bottom') +
  ylab('') +
  xlab('Median Predicted Bloom Date') +
  guides(fill=ggplot2::guide_legend(title="Weather Scenario"))
ggplot2::ggsave('figures/paper/change_med_bloom_2050_v2_updated_gcms.jpeg', device = 'jpeg',
                height = 40, width = 31, units = 'cm')


shift_df %>% 
  mutate(dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         location = factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax')),
         spec = as.numeric(factor(species, levels = c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))),
         loc = as.numeric(factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax'))),
         med_future = ifelse(is.na(med_current), yes = NA, no = med_future)) %>% 
  filter(scenario_year == '2085') %>% 
  ggplot(ggplot2::aes(y = spec)) +
  geom_rect(aes(xmin = med_current, xmax = med_future, ymin = spec + dodge_low, ymax = spec + dodge_up, fill = ssp)) +
  geom_point(aes(x = med_future, y = spec + ((dodge_low + dodge_up)/2)),
             col = 'black',
             show.legend = FALSE, shape = 4, size = 1) + 
  geom_rect(aes(xmin = med_current - 0.5, xmax = med_current + 0.5, ymax = spec - 0.4, ymin = spec + 0.4, fill = 'Simulation 2015'),  size = 2) +
  # geom_bar(stat = 'identity', position = 'dodge') +
  facet_grid(location~., scales = 'free_y', space = 'free_y') +
  scale_color_manual(values = c("#56B4E9", "#009E73","#F0E442",  "#E69F00"))+
  scale_fill_manual(values = c('black', "#56B4E9", "#009E73","#F0E442",  "#E69F00", 'grey70'))+
  theme_bw(base_size = 15) +
  scale_x_continuous(breaks = c(32,  60, 91, 121, 152), 
                     labels = c('Feb', 'Mar', 'Apr', 'May', 'Jun'),
                     minor_breaks = c(1, 32, 60, 91, 121, 152)) +
  scale_y_reverse(breaks = rep(1:7,6), labels = rep(c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'), 6)) +
  coord_cartesian(xlim = c(32, 155)) +
  theme(axis.text.x = ggplot2::element_text(angle = 45, hjust = 1), legend.position = 'bottom') +
  ylab('') +
  xlab('Median Predicted Bloom Date') +
  guides(fill=ggplot2::guide_legend(title="Weather Scenario"))
ggplot2::ggsave('figures/paper/change_med_bloom_2085_v2_updated_gcms.jpeg', device = 'jpeg',
                height = 40, width = 31, units = 'cm')




#similar plot but now comparing local fitted vs foreign projections of species


shift_df_locfor <- merge(median_2015_locfor, median_future_local_vs_foreign, by = c('species','location', 'local_cultivar')) %>% 
  mutate(shift_bloom = round(med_future - med_current, digits = 2))

shift_table <- shift_df_locfor %>% 
  group_by(species, location, scenario_year, ssp, local_cultivar) %>% 
  summarise(min_shift = min(shift_bloom, na.rm = TRUE),
            max_shift = max(shift_bloom, na.rm = TRUE))

shift_table %>% 
  ggplot(aes(x = local_cultivar, y= med_shift)) +
  geom_boxplot() +
  facet_grid(species ~ location)

shift_df_locfor %>% 
  mutate(dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         location = factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax')),
         spec = as.numeric(factor(species, levels = c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))),
         loc = as.numeric(factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax'))),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
         local_cultivar = factor(local_cultivar, levels = c(TRUE, FALSE), labels = c('local calibration', 'projected')),
         med_future = ifelse(is.na(med_current), yes = NA, no = med_future)) %>% 
  drop_na() %>% 
  ggplot(aes(x = local_cultivar, y= shift_bloom, fill = location)) +
  geom_boxplot() 
    

shift_df_locfor %>% 
  mutate(dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         location = factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax')),
         spec = as.numeric(factor(species, levels = c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))),
         loc = as.numeric(factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax'))),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
         local_cultivar = factor(local_cultivar, levels = c(TRUE, FALSE), labels = c('local calibration', 'projected')),
         med_future = ifelse(is.na(med_current), yes = NA, no = med_future)) %>% 
  drop_na() %>% 
  filter(scenario_year == '2050') %>% 
  ggplot(ggplot2::aes(y = spec_loc)) +
  geom_rect(aes(xmin = med_current, xmax = med_future, ymin = spec_loc + dodge_low, ymax = spec_loc + dodge_up, fill = ssp), alpha = 0.2) +
  geom_point(aes(x = med_future, y = spec_loc + ((dodge_low + dodge_up)/2)),
             col = 'black',
             show.legend = FALSE, shape = 4, size = 1) + 
  geom_rect(aes(xmin = med_current - 0.5, xmax = med_current + 0.5, ymax = spec_loc - 0.4, ymin = spec_loc + 0.4, fill = 'Simulation 2015'),  size = 2) +
  # geom_bar(stat = 'identity', position = 'dodge') +
  facet_grid(species~local_cultivar, scales = 'free_y', space = 'free_y') +
  scale_color_manual(values = c("#56B4E9", "#009E73","#F0E442",  "#E69F00"))+
  scale_fill_manual(values = c('black', "#56B4E9", "#009E73","#F0E442",  "#E69F00", 'grey70'))+
  theme_bw(base_size = 15) +
  scale_x_continuous(breaks = c(32,  60, 91, 121, 152), 
                     labels = c('Feb', 'Mar', 'Apr', 'May', 'Jun'),
                     minor_breaks = c(1, 32, 60, 91, 121, 152)) +
  scale_y_reverse(breaks = 1:13, labels = c('Klein-Altendorf', 'Meknes', 'Klein-Altendorf', 'Zaragoza', 'Zaragoza', 'Cieza', 'Klein-Altendorf', 'Klein-Altendorf', 'Zaragoza', 'Santomera', 'Meknes', 'Sfax', 'Sfax')) +
  coord_cartesian(xlim = c(32, 155)) +
  theme(axis.text.x = ggplot2::element_text(angle = 45, hjust = 1), legend.position = 'bottom') +
  ylab('') +
  xlab('Median Predicted Bloom Date') +
  guides(fill=ggplot2::guide_legend(title="Weather Scenario"))
ggplot2::ggsave('figures/paper/change_med_bloom_2050_loc-vs-for_v3_updated_gcms.jpeg', device = 'jpeg',
                height = 40, width = 31, units = 'cm')


#data.frame for heat map of median unfulfilled thermal requirements
fail_plot <- fail_summarized %>% 
  mutate(species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
         local_cultivar = factor(local_cultivar, levels = c(TRUE, FALSE), labels = c('Local calibration', 'Projected')),
         dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         median_fail_text = round(median_fail, digits = 0),
         #median_fail_text = ifelse(median_fail < 10, yes = paste0(' ', median_fail_text), no = median_fail_text),
         median_fail_text = ifelse(median_fail < 1 & median_fail > 0, yes = '<1', no = median_fail_text),
         median_fail_text = ifelse(median_fail < 100 & median_fail > 99, yes = '>99', no = median_fail_text)) %>%
  #mutate(median_fail_text = paste0(median_fail_text, ' %')) %>% 
  drop_na() 

#decides where to start the heatmap
x_cutoff <- 157
heat_map_width <- 11.6


#data.frame controlling the positioning of the "No Pred." label in the plot
df_text_no_pred <- data.frame(y =  rep(c(0.5,2.5, 4.5, 6.5, 7.5, 9.45, 12.5), 2),
                              x = x_cutoff + 0.5,
                              species = factor(rep(c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio'), 2),
                                               levels =  c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
                              local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 7), levels = c('Local calibration', 'Projected')))

df_padding_box <- data.frame(y = rep(c(6.45, 12.45), 2),
                             x = x_cutoff +1,
                             species = factor(rep(c('Europ. Plum', 'Pistachio'), 2),
                                              levels =  c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
                             local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 2), levels = c('Local calibration', 'Projected')))


#data.frame to control the grey box background of the heatmap
grey_box_df <- data.frame(spec_loc = NA,
           start = 152,
           end = Inf,
           sepecies = factor(rep(c('Apple', 'Apple',  'Pear',  'Pear', 'Apricot', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Sweet Cherry', rep('Almond', 3), 'Pistachio'), 2),
                             levels =  c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
           local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 13), levels = c('Local calibration', 'Projected')))


#needed for subsetting the data.frame
year_used <- 2050

shift_df_locfor %>% 
  mutate(dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         location = factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax')),
         spec = as.numeric(factor(species, levels = c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))),
         loc = as.numeric(factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax'))),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         ssp_label = factor(ssp, levels = c('ssp126', 'ssp245', 'ssp370', 'ssp585'), labels = c('SSP1', 'SSP2', 'SSP3', 'SSP5')),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
         local_cultivar = factor(local_cultivar, levels = c(TRUE, FALSE), labels = c('Local calibration', 'Projected')),
         med_future = ifelse(is.na(med_current), yes = NA, no = med_future)) %>% 
  drop_na() %>% 
  filter(scenario_year == as.character(year_used)) %>% 
  ggplot(ggplot2::aes(y = spec_loc)) +
  geom_rect(aes(xmin = med_current, xmax = med_future, ymin = spec_loc + dodge_low, ymax = spec_loc + dodge_up, fill = ssp_label), alpha = 0.2) +
  geom_point(aes(x = med_future, y = spec_loc + ((dodge_low + dodge_up)/2), 
                 shape = 'Median prediction\nof individual GCMs'),
             col = 'black',
             show.legend = TRUE, size = 1) + 
  geom_rect(aes(xmin = med_current - 0.5, xmax = med_current + 0.5, ymax = spec_loc - 0.4, ymin = spec_loc + 0.4, fill = 'Simulation 2015'),  size = 2) +
  geom_rect(data = grey_box_df, xmin = x_cutoff, xmax = Inf,ymax = Inf, ymin = -Inf, fill = 'grey85',  size = 2) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+3, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+5, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+7, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+9, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+11, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+13, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+15, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  scale_colour_gradient(low = "white", high = "firebrick2", limits = c(0, 100),name = 'Median Failure Rate') +
  geom_vline(xintercept = x_cutoff) +
  geom_point(data = df_padding_box, aes(y = y, x = x), col = 'grey85', size = 0.1) +
  geom_text(data = df_text_no_pred, aes(y = y, x = x,
                                        label = 'No Pred.'),hjust = 0) +
  # geom_text(data = fail_plot[fail_plot$scenario_year == year_used],aes(x = 155 + 2, y = spec_loc  + ((dodge_low + dodge_up)/2), 
  #               label = paste(format(round(median_fail, digits = 0), nsmall = 0), '%')), size = 3) +
  geom_text(data = fail_plot[fail_plot$scenario_year == year_used],aes(x = x_cutoff + heat_map_width * 1.2, y = spec_loc  + ((dodge_low + dodge_up)/2),
                label = median_fail_text), size = 3, hjust = 1) +
  facet_grid(species~local_cultivar, scales = 'free_y', space = 'free_y') +
  #scale_color_manual(values = c("#56B4E9", "#009E73","#F0E442",  "#E69F00"))+
  scale_fill_manual(breaks = c('SSP1', 'SSP2',  'Simulation 2015', 'SSP3', 'SSP5'),  
                    values = c("#56B4E9", "#009E73",'black',"#F0E442",  "#E69F00", 'grey70'), name = "Weather Scenario")+
  scale_shape_manual(values = 4, name = NULL) + 
  theme_bw(base_size = 15) +
  scale_x_continuous(breaks = c(32,  60, 91, 121, 152), 
                     labels = c('Feb', 'Mar', 'Apr', 'May', 'Jun'),
                     minor_breaks = c(1, 32, 60, 91, 121, 152)) +
  scale_y_reverse(breaks = 1:13, labels = c('Klein-Altendorf', 'Meknes', 'Klein-Altendorf', 'Zaragoza', 'Zaragoza', 'Cieza', 'Klein-Altendorf', 'Klein-Altendorf', 'Zaragoza', 'Santomera', 'Meknes', 'Sfax', 'Sfax'),
                  minor_breaks = NULL) +
  coord_cartesian(xlim = c(32, x_cutoff + heat_map_width)) +
  guides( fill=guide_legend(nrow=2,byrow=TRUE, order = 1)) +
  theme(legend.position = 'bottom') +
  ylab('') +
  xlab('Median Predicted Bloom Date')
ggplot2::ggsave('figures/paper/change_med_bloom_2050_loc-vs-for_v4_updated_gcms.jpeg', device = 'jpeg',
                height = 42, width = 31, units = 'cm')




fail_plot <- fail_plot %>% 
  mutate(species = factor(species, levels = c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Almond', 'Pist.')))

#data.frame controlling the positioning of the "No Pred." label in the plot
df_text_no_pred <- data.frame(y =  rep(c(0.45,2.45, 4.45, 6.45, 7.45, 9.475, 12.45), 2),
                              x = x_cutoff + 1,
                              species = factor(rep(c('Apple', 'Pear', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Almond', 'Pist.'), 2),
                                               levels =  c('Apple', 'Pear', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Almond', 'Pist.')),
                              local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 7), levels = c('Local calibration', 'Projected')))

df_padding_box <- data.frame(y = rep(c(6.4, 12.4), 2),
                             x = x_cutoff +1,
                             species = factor(rep(c('E. Plum', 'Pist.'), 2),
                                              levels =  c('Apple', 'Pear', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Almond', 'Pist.')),
                             local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 2), levels = c('Local calibration', 'Projected')))

df_padding_box$species


#data.frame to control the grey box background of the heatmap
grey_box_df <- data.frame(spec_loc = NA,
                          start = 152,
                          end = Inf,
                          sepecies = factor(rep(c('Apple', 'Apple',  'Pear',  'Pear', 'Apricot', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Sweet Cherry', rep('Almond', 3), 'Pist.'), 2),
                                            levels =  c('Apple', 'Pear', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Almond', 'Pist.')),
                          local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 13), levels = c('Local calibration', 'Projected')))

box_size <-  4

p <- shift_df_locfor %>% 
  mutate(dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         location = factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax')),
         spec = as.numeric(factor(species, levels = c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))),
         loc = as.numeric(factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax'))),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         ssp_label = factor(ssp, levels = c('ssp126', 'ssp245', 'ssp370', 'ssp585'), labels = c('SSP1', 'SSP2', 'SSP3', 'SSP5')),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Almond', 'Pist.')),
         local_cultivar = factor(local_cultivar, levels = c(TRUE, FALSE), labels = c('Local calibration', 'Projected')),
         med_future = ifelse(is.na(med_current), yes = NA, no = med_future)) %>% 
  drop_na() %>% 
  filter(scenario_year == as.character(year_used)) %>% 
  ggplot(ggplot2::aes(y = spec_loc)) +
  geom_rect(aes(xmin = med_current, xmax = med_future, ymin = spec_loc + dodge_low, ymax = spec_loc + dodge_up, fill = ssp_label), alpha = 0.2) +
  # geom_point(aes(x = med_future, y = spec_loc + ((dodge_low + dodge_up)/2), 
  #                shape = 'Median prediction\nof individual GCMs'),
  #            col = 'grey20',
  #            show.legend = TRUE, size = 0.7) + 
  geom_rect(aes(xmin = med_current - 0.5, xmax = med_current + 0.5, ymax = spec_loc - 0.4, ymin = spec_loc + 0.4, fill = 'Simulation 2015'),  size = 2) +
  geom_rect(data = grey_box_df, xmin = x_cutoff, xmax = Inf,ymax = Inf, ymin = -Inf, fill = 'grey85',  size = 2) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+4.5, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+5, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+7, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+9, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+11, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+13, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+15, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=box_size) +
  scale_colour_gradient(low = "white", high = "firebrick2", limits = c(0, 100),name = 'Median Failure Rate') +
  geom_vline(xintercept = x_cutoff) +
  geom_point(data = df_padding_box, aes(y = y, x = x), col = 'grey85', size = 0.1) +
  geom_text(data = df_text_no_pred, aes(y = y, x = x,
                                        label = 'Fail'),hjust = 0, size = 2.5) +
  # geom_text(data = fail_plot[fail_plot$scenario_year == year_used],aes(x = 155 + 2, y = spec_loc  + ((dodge_low + dodge_up)/2), 
  #               label = paste(format(round(median_fail, digits = 0), nsmall = 0), '%')), size = 3) +
  geom_text(data = fail_plot[fail_plot$scenario_year == year_used],aes(x = x_cutoff + (heat_map_width * 1.5), y = spec_loc  + ((dodge_low + dodge_up)/2),
                                                                       label = median_fail_text), size = 2.5, hjust = 1) +
  facet_grid(species~local_cultivar, scales = 'free_y', space = 'free_y') +
  #scale_color_manual(values = c("#56B4E9", "#009E73","#F0E442",  "#E69F00"))+
  scale_fill_manual(breaks = c('SSP1', 'SSP2',  'Simulation 2015', 'SSP3', 'SSP5'),  
                    values = c("#56B4E9", "#009E73",'black',"#F0E442",  "#E69F00", 'grey70'), name = "Weather Scenario")+
  #scale_shape_manual(values = 4, name = NULL) + 
  theme_bw(base_size = 12) +
  scale_x_continuous(breaks = c(32,  60, 91, 121, 152), 
                     labels = c('Feb', 'Mar', 'Apr', 'May', 'Jun'),
                     minor_breaks = c(1, 32, 60, 91, 121, 152)) +
  scale_y_reverse(breaks = 1:13, labels = c('Klein-Altendorf', 'Meknes', 'Klein-Altendorf', 'Zaragoza', 'Zaragoza', 'Cieza', 'Klein-Altendorf', 'Klein-Altendorf', 'Zaragoza', 'Santomera', 'Meknes', 'Sfax', 'Sfax'),
                  minor_breaks = NULL) +
  coord_cartesian(xlim = c(32, x_cutoff + heat_map_width)) +
  guides( fill=guide_legend(nrow=2,byrow=TRUE, order = 1)) +
  theme(legend.position = 'bottom') +
  ylab('') +
  xlab('Median Predicted Bloom Date')

jpeg(file="figures/paper/legend_shift.jpeg", width = 25, height = 5, units = 'cm', res = 200)
leg <- cowplot::get_legend(p)
grid::grid.newpage()
grid::grid.draw(leg)
dev.off()

p+
  theme(legend.position = "none")
ggplot2::ggsave('figures/paper/change_med_bloom_2050_loc-vs-for_v4_updated_gcms_resized.jpeg', device = 'jpeg',
                height = 20, width = 16, units = 'cm')



#data.frame controlling the positioning of the "No Pred." label in the plot
df_text_no_pred <- data.frame(y =  rep(c(0.45,2.45, 4.45, 6.45, 7.45, 9.46, 12.45), 2),
                              x = x_cutoff + 2.5,
                              species = factor(rep(c('Apple', 'Pear', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Almond', 'Pist.'), 2),
                                               levels =  c('Apple', 'Pear', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Almond', 'Pist.')),
                              local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 7), levels = c('Local calibration', 'Projected')),
                              scenario_year = rep(c(2050, 2085), each = 7))

df_padding_box <- data.frame(y = rep(c(6.4, 12.4), 2),
                             x = x_cutoff +1,
                             species = factor(rep(c('E. Plum', 'Pist.'), 2),
                                              levels =  c('Apple', 'Pear', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Almond', 'Pist.')),
                             local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 2), levels = c('Local calibration', 'Projected')),
                             scenario_year = rep(c(2050, 2085), each = 2))

df_padding_box$species


#data.frame to control the grey box background of the heatmap
grey_box_df <- data.frame(spec_loc = NA,
                          start = 152,
                          end = Inf,
                          sepecies = factor(rep(c('Apple', 'Apple',  'Pear',  'Pear', 'Apricot', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Sweet Cherry', rep('Almond', 3), 'Pist.'), 2),
                                            levels =  c('Apple', 'Pear', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Almond', 'Pist.')),
                          local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 13), levels = c('Local calibration', 'Projected')),
                          scenario_year = rep(c(2050, 2085), each = 13))

library(colorRamps)

box_size <-  3

p <- shift_df_locfor %>% 
  mutate(dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         location = factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax')),
         spec = as.numeric(factor(species, levels = c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))),
         loc = as.numeric(factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax'))),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         ssp_label = factor(ssp, levels = c('ssp126', 'ssp245', 'ssp370', 'ssp585'), labels = c('SSP1', 'SSP2', 'SSP3', 'SSP5')),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Almond', 'Pist.')),
         local_cultivar = factor(local_cultivar, levels = c(TRUE, FALSE), labels = c('Local calibration', 'Projected')),
         med_future = ifelse(is.na(med_current), yes = NA, no = med_future)) %>% 
  drop_na() %>% 
  filter(local_cultivar == 'Local calibration') %>% 
  ggplot(ggplot2::aes(y = spec_loc)) +
  geom_rect(aes(xmin = med_current, xmax = med_future, ymin = spec_loc + dodge_low, ymax = spec_loc + dodge_up, fill = ssp_label), alpha = 0.2) +
  # geom_point(aes(x = med_future, y = spec_loc + ((dodge_low + dodge_up)/2), 
  #                shape = 'Median prediction\nof individual GCMs'),
  #            col = 'grey20',
  #            show.legend = TRUE, size = 0.7) + 
  geom_rect(aes(xmin = med_current - 0.5, xmax = med_current + 0.5, ymax = spec_loc - 0.4, ymin = spec_loc + 0.4, fill = 'Simulation 2015'),  size = 2) +
  geom_rect(data = grey_box_df, xmin = x_cutoff, xmax = Inf,ymax = Inf, ymin = -Inf, fill = 'grey85',  size = 2) +
  geom_point(data = fail_plot[fail_plot$local_cultivar == 'Local calibration'], aes(x=x_cutoff+3.7, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail/100), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$local_cultivar == 'Local calibration'], aes(x=x_cutoff+4.5, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail/100), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$local_cultivar == 'Local calibration'], aes(x=x_cutoff+5, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail/100), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$local_cultivar == 'Local calibration'], aes(x=x_cutoff+7, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail/100), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$local_cultivar == 'Local calibration'], aes(x=x_cutoff+9, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail/100), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$local_cultivar == 'Local calibration'], aes(x=x_cutoff+11, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail/100), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$local_cultivar == 'Local calibration'], aes(x=x_cutoff+13, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail/100), shape=15, size=box_size) +
  geom_point(data = fail_plot[fail_plot$local_cultivar == 'Local calibration'], aes(x=x_cutoff+15, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail/100), shape=15, size=box_size) +
  #scale_colour_gradient(low = "white", high = "firebrick2", limits = c(0, 100),name = 'Median Failure Rate', labels = scales::percent) +
  scale_color_gradientn(colours = matlab.like(15),
                       labels = scales::percent,
                       limits = c(0, 1),
                       name = paste0('Median Failure Rate'))+
  geom_vline(xintercept = x_cutoff) +
  geom_point(data = df_padding_box, aes(y = y, x = x), col = 'grey85', size = 0.1) +
  geom_text(data = df_text_no_pred, aes(y = y, x = x,
                                        label = 'Fail'),hjust = 0, size = 2.5) +
  # geom_text(data = fail_plot[fail_plot$scenario_year == year_used],aes(x = 155 + 2, y = spec_loc  + ((dodge_low + dodge_up)/2), 
  #               label = paste(format(round(median_fail, digits = 0), nsmall = 0), '%')), size = 3) +
  geom_text(data = fail_plot[fail_plot$local_cultivar == 'Local calibration'],aes(x = x_cutoff + (heat_map_width * 1.1), y = spec_loc  + ((dodge_low + dodge_up)/2),
                                                                       label = median_fail_text), size = 2.5, hjust = 1) +
  facet_grid(species~scenario_year, scales = 'free_y', space = 'free_y') +
  #scale_color_manual(values = c("#56B4E9", "#009E73","#F0E442",  "#E69F00"))+
  scale_fill_manual(breaks = c('SSP1', 'SSP2',  'Simulation 2015', 'SSP3', 'SSP5'),  
                    values = c("#56B4E9", "#009E73",'black',"#F0E442",  "#E69F00", 'grey70'), name = "Weather Scenario")+
  #scale_shape_manual(values = 4, name = NULL) + 
  theme_bw(base_size = 12) +
  scale_x_continuous(breaks = c(32,  60, 91, 121, 152), 
                     labels = c('Feb', 'Mar', 'Apr', 'May', 'Jun'),
                     minor_breaks = c(1, 32, 60, 91, 121, 152)) +
  scale_y_reverse(breaks = 1:13, labels = c('Klein-Altendorf', 'Meknes', 'Klein-Altendorf', 'Zaragoza', 'Zaragoza', 'Cieza', 'Klein-Altendorf', 'Klein-Altendorf', 'Zaragoza', 'Santomera', 'Meknes', 'Sfax', 'Sfax'),
                  minor_breaks = NULL) +
  coord_cartesian(xlim = c(32, x_cutoff + heat_map_width)) +
  guides( fill=guide_legend(nrow=2,byrow=TRUE, order = 1)) +
  theme(legend.position = 'bottom') +
  ylab('') +
  xlab('Median Predicted Bloom Date')

jpeg(file="figures/paper/legend_shift.jpeg", width = 25, height = 5, units = 'cm', res = 200)
leg <- cowplot::get_legend(p)
grid::grid.newpage()
grid::grid.draw(leg)
dev.off()

p+
  theme(legend.position = "none")
ggplot2::ggsave('figures/paper/change_med_bloom_2050_loc-vs-for_v4_updated_gcms_only-local.jpeg', device = 'jpeg',
                height = 20, width = 16, units = 'cm')


cm_to_inch <- 1/2.54
svg('figures/paper/change_med_bloom_only_loc.svg', width = 16 * cm_to_inch, height = 20*cm_to_inch)
p+
  theme(legend.position = "none")
dev.off()























df_shift <- shift_df_locfor %>% 
  mutate(dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         location = factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax')),
         spec = as.numeric(factor(species, levels = c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))),
         loc = as.numeric(factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax'))),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         ssp_label = factor(ssp, levels = c('ssp126', 'ssp245', 'ssp370', 'ssp585'), labels = c('SSP1', 'SSP2', 'SSP3', 'SSP5')),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'E. Plum', 'Sweet Cherry', 'Almond', 'Pist.')),
         local_cultivar = factor(local_cultivar, levels = c(TRUE, FALSE), labels = c('Local calibration', 'Projected')),
         med_future = ifelse(is.na(med_current), yes = NA, no = med_future)) %>% 
  drop_na() %>% 
  filter(local_cultivar == 'Local calibration') %>% 
  group_by(species, location, scenario_year, ssp) %>% 
  summarize(med_shift = median(med_current - med_future))










year_used <- 2085
shift_df_locfor %>% 
  mutate(dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         location = factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax')),
         spec = as.numeric(factor(species, levels = c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))),
         loc = as.numeric(factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax'))),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         ssp_label = factor(ssp, levels = c('ssp126', 'ssp245', 'ssp370', 'ssp585'), labels = c('SSP1', 'SSP2', 'SSP3', 'SSP5')),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
         local_cultivar = factor(local_cultivar, levels = c(TRUE, FALSE), labels = c('Local calibration', 'Projected')),
         med_future = ifelse(is.na(med_current), yes = NA, no = med_future)) %>% 
  drop_na() %>% 
  filter(scenario_year == as.character(year_used)) %>% 
  ggplot(ggplot2::aes(y = spec_loc)) +
  geom_rect(aes(xmin = med_current, xmax = med_future, ymin = spec_loc + dodge_low, ymax = spec_loc + dodge_up, fill = ssp_label), alpha = 0.2) +
  geom_point(aes(x = med_future, y = spec_loc + ((dodge_low + dodge_up)/2), 
                 shape = 'Median prediction\nof individual GCMs'),
             col = 'black',
             show.legend = TRUE, size = 1) + 
  geom_rect(aes(xmin = med_current - 0.5, xmax = med_current + 0.5, ymax = spec_loc - 0.4, ymin = spec_loc + 0.4, fill = 'Simulation 2015'),  size = 2) +
  geom_rect(data = grey_box_df, xmin = x_cutoff, xmax = Inf,ymax = Inf, ymin = -Inf, fill = 'grey85',  size = 2) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+3, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+5, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+7, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+9, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+11, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+13, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = fail_plot[fail_plot$scenario_year == year_used], aes(x=x_cutoff+15, y= spec_loc + ((dodge_low + dodge_up)/2), color=median_fail), shape=15, size=7.5) +
  geom_point(data = df_padding_box, aes(y = y, x = x), col = 'grey85', size = 0.1) +
  scale_colour_gradient(low = "white", high = "firebrick2", limits = c(0, 100),name = 'Median Failure Rate') +
  geom_vline(xintercept = x_cutoff) +
  geom_text(data = df_text_no_pred, aes(y = y, x = x,
                                        label = 'No Pred.'),hjust = 0) +
  # geom_text(data = fail_plot[fail_plot$scenario_year == year_used],aes(x = 155 + 2, y = spec_loc  + ((dodge_low + dodge_up)/2), 
  #               label = paste(format(round(median_fail, digits = 0), nsmall = 0), '%')), size = 3) +
  geom_text(data = fail_plot[fail_plot$scenario_year == year_used],aes(x = x_cutoff + heat_map_width * 1.2, y = spec_loc  + ((dodge_low + dodge_up)/2),
                                                                       label = median_fail_text), size = 3, hjust = 1) +
  facet_grid(species~local_cultivar, scales = 'free_y', space = 'free_y') +
  #scale_color_manual(values = c("#56B4E9", "#009E73","#F0E442",  "#E69F00"))+
  scale_fill_manual(breaks = c('SSP1', 'SSP2',  'Simulation 2015', 'SSP3', 'SSP5'),  
                    values = c("#56B4E9", "#009E73",'black',"#F0E442",  "#E69F00", 'grey70'), name = "Weather Scenario")+
  scale_shape_manual(values = 4, name = NULL) + 
  theme_bw(base_size = 15) +
  scale_x_continuous(breaks = c(32,  60, 91, 121, 152), 
                     labels = c('Feb', 'Mar', 'Apr', 'May', 'Jun'),
                     minor_breaks = c(1, 32, 60, 91, 121, 152)) +
  scale_y_reverse(breaks = 1:13, 
                  labels = c('Klein-Altendorf', 'Meknes', 'Klein-Altendorf', 'Zaragoza', 'Zaragoza', 'Cieza', 'Klein-Altendorf', 'Klein-Altendorf', 'Zaragoza', 'Santomera', 'Meknes', 'Sfax', 'Sfax'), 
                  minor_breaks = NULL) +
  guides( fill=guide_legend(nrow=2,byrow=TRUE, order = 1)) +
  coord_cartesian(xlim = c(32, x_cutoff + heat_map_width)) +
  theme(legend.position = 'bottom') +
  ylab('') +
  xlab('Median Predicted Bloom Date')
ggplot2::ggsave('figures/paper/change_med_bloom_2085_loc-vs-for_v4_updated_gcms.jpeg', device = 'jpeg',
                height = 42, width = 31, units = 'cm')




shift_table <- shift_df_locfor %>% 
  mutate(dodge_up = recode(ssp, ssp126 = -0.2, ssp245 = 0, ssp370 = 0.2, ssp585 = 0.4),
         dodge_low = recode(ssp, ssp126 = -0.4, ssp245 = -0.2, ssp370 = 0, ssp585 = 0.2),
         location = factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax')),
         spec = as.numeric(factor(species, levels = c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))),
         loc = as.numeric(factor(location, levels = c('Klein-Altendorf', 'Zaragoza', 'Cieza', 'Santomera', 'Meknes', 'Sfax'))),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         ssp_label = factor(ssp, levels = c('ssp126', 'ssp245', 'ssp370', 'ssp585'), labels = c('SSP1', 'SSP2', 'SSP3', 'SSP5')),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
         local_cultivar = factor(local_cultivar, levels = c(TRUE, FALSE), labels = c('Local calibration', 'Projected')),
         med_future = ifelse(is.na(med_current), yes = NA, no = med_future)) %>% 
  drop_na() %>% 
  group_by(species, location, scenario_year, ssp, local_cultivar) %>% 
  summarize(min_shift = min(shift_bloom),
            med_shift = median(shift_bloom),
            max_shift = max(shift_bloom))





#makes similar plot for 2015 (but maybe without the flowering time window)

fail_2015 <- pheno_2015 %>% 
  mutate(spec_cult_loc = paste(species, cultivar, location, sep = "_"),
         local_cultivar = ifelse(spec_cult_loc %in% master_fitting_data$spec_cult_loc, yes = 'Local calibration', no = 'Projected'),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                           "Apple_Meknes",
                                                                           "Pear_Klein-Altendorf",
                                                                           "Pear_Zaragoza",
                                                                           "Apricot_Zaragoza",
                                                                           "Apricot_Cieza",
                                                                           "European Plum_Klein-Altendorf",
                                                                           "Sweet Cherry_Klein-Altendorf",
                                                                           "Sweet Cherry_Zaragoza",
                                                                           "Almond_Santomera",
                                                                           "Almond_Meknes",
                                                                           "Almond_Sfax",
                                                                           "Pistachio_Sfax")),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
         fail = is.na(pheno_predicted)) %>%
  select(species, location, local_cultivar, fail, spec_loc) %>% 
  drop_na() %>% 
  group_by(species, location, local_cultivar, spec_loc) %>% 
  summarize(fail_rate = (sum(fail) / n()) * 100) %>% 
  mutate(fail_text = round(fail_rate, digits = 0),
         fail_text = ifelse(fail_rate < 1 & fail_rate > 0, yes = '<1', no = as.character(fail_text)),
         fail_text = ifelse(fail_rate > 99 & fail_rate < 100, yes = '>99', no = fail_text),
         fail_text = paste0(fail_text, ' %'))

x_cutoff <- 162
point_size <- 15
heat_map_width <- 25

df_text_no_pred <- data.frame(y =  rep(c(0.5,2.5, 4.5, 6.5, 7.5, 9.45, 12.5), 2),
                              x = x_cutoff + 0.5,
                              species = factor(rep(c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio'), 2),
                                               levels =  c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
                              local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 7), levels = c('Local calibration', 'Projected')))

df_padding_box <- data.frame(y = rep(c(6.45, 12.45, 7.55, 13.55), 2),
                             x = x_cutoff +1,
                             species = factor(rep(c('Europ. Plum', 'Pistachio'), 4),
                                              levels =  c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
                             local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 4), levels = c('Local calibration', 'Projected')))


pheno_2015 %>% 
  mutate(spec_cult_loc = paste(species, cultivar, location, sep = "_"),
         local_cultivar = ifelse(spec_cult_loc %in% master_fitting_data$spec_cult_loc, yes = 'Local calibration', no = 'Projected'),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))) %>% 
  drop_na() %>% 
  ggplot(ggplot2::aes(y = spec_loc)) +
  geom_boxplot(aes(y = spec_loc, x = pheno_predicted, group = spec_loc, fill = species), width = 0.6, show.legend = FALSE) +
  geom_rect(data = grey_box_df, xmin = x_cutoff, xmax = Inf,ymax = Inf, ymin = -Inf, fill = 'grey85',  size = 2) +
  #geom_point(data = fail_2015, aes(x=x_cutoff+3, y= spec_loc, color=fail_rate), shape=15, size=point_size) +
  geom_point(data = fail_2015, aes(x=x_cutoff+10+1, y= spec_loc, color=fail_rate), shape=15, size=point_size) +
  geom_point(data = fail_2015, aes(x=x_cutoff+12, y= spec_loc, color=fail_rate), shape=15, size=point_size) +
  geom_point(data = fail_2015, aes(x=x_cutoff+14, y= spec_loc, color=fail_rate), shape=15, size=point_size) +
  geom_point(data = fail_2015, aes(x=x_cutoff+16, y= spec_loc, color=fail_rate), shape=15, size=point_size) +
  geom_point(data = fail_2015, aes(x=x_cutoff+18, y= spec_loc, color=fail_rate), shape=15, size=point_size) +
  geom_point(data = fail_2015, aes(x=x_cutoff+20, y= spec_loc, color=fail_rate), shape=15, size=point_size) +
  geom_point(data = fail_2015, aes(x=x_cutoff+22, y= spec_loc, color=fail_rate), shape=15, size=point_size) +
  geom_point(data = df_padding_box, aes(y = y, x = x), col = 'grey85', size = 0.1) +
  scale_colour_gradient(low = "white", high = "firebrick2", limits = c(0, 100),name = 'Failure Rate') +
  geom_vline(xintercept = x_cutoff) +
  geom_text(data = df_text_no_pred, aes(y = y, x = x,
                                        label = 'No Pred.'),hjust = 0) +
  # geom_text(data = fail_plot[fail_plot$scenario_year == year_used],aes(x = 155 + 2, y = spec_loc  + ((dodge_low + dodge_up)/2), 
  #               label = paste(format(round(median_fail, digits = 0), nsmall = 0), '%')), size = 3) +
  geom_text(data = fail_2015,aes(x = x_cutoff + (heat_map_width + (heat_map_width/10)), y = spec_loc,
                                                                       label = fail_text), hjust = 1) +
  facet_grid(species~local_cultivar, scales = 'free_y', space = 'free_y') +
  #scale_color_manual(values = c("#56B4E9", "#009E73","#F0E442",  "#E69F00"))+
  scale_fill_manual(values = c("#009E73", "#009E73", "#E69F00","#E69F00" ,"#E69F00", "#56B4E9", "#56B4E9" ), name = NULL)+
  scale_shape_manual(values = 4, name = NULL) + 
  theme_bw(base_size = 15) +
  scale_x_continuous(breaks = c(1, 32,  60, 91, 121, 152), 
                     labels = c('Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun'),
                     minor_breaks = c(1, 32, 60, 91, 121, 152)) +
  scale_y_reverse(breaks = 1:13, 
                  labels = c('Klein-Altendorf', 'Meknes', 'Klein-Altendorf', 'Zaragoza', 'Zaragoza', 'Cieza', 'Klein-Altendorf', 'Klein-Altendorf', 'Zaragoza', 'Santomera', 'Meknes', 'Sfax', 'Sfax'), 
                  minor_breaks = NULL) +
  guides( fill=guide_legend(nrow=2,byrow=TRUE, order = 1)) +
  coord_cartesian(xlim = c(0, x_cutoff + heat_map_width)) +
  theme(legend.position = 'bottom') +
  ylab('') +
  xlab('Predicted Bloom Date')
ggsave('figures/paper/2015_predictions_local-vs-foreign.jpeg', height = 30, width = 25, units = 'cm', device = 'jpeg')
  


pheno_2015 %>% 
  mutate(spec_cult_loc = paste(species, cultivar, location, sep = "_"),
         local_cultivar = ifelse(spec_cult_loc %in% master_fitting_data$spec_cult_loc, yes = 'Local calibration', no = 'Projected'),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))) %>% 
  drop_na() %>% 
  filter(species == 'Almond') %>% 
  group_by(location, local_cultivar) %>% 
  summarize(med = median(pheno_predicted))
           


median_future_local_vs_foreign_allgcm

#so why do the predictions of local calibrated and "foreign" cultivars differ? what makes them different?
median_2015_locfor %>% 
  mutate(ssp = '2015', 
         scenario_year = 2015) %>% 
  rename(`med_future` = 'med_current',
         `sd_future` = 'sd_current') %>% 
  rbind(median_future_local_vs_foreign_allgcm) %>% 
  mutate(local_cultivar = factor(local_cultivar, levels = c(TRUE, FALSE), labels = c('Local calibration', 'Projected')),
         ssp_local = paste(ssp, local_cultivar, sep = '_'), 
         dodge = recode(ssp_local, 
                        `2015_Projected` = 0.1,
                        `2015_Local calibration` = -0.1,
                        `ssp126_Projected` = 0.25,
                        `ssp126_Local calibration` =0.2,
                        `ssp245_Projected` = 0.1,
                        `ssp245_Local calibration` = 0.05,
                        `ssp370_Projected` = -0.05,
                        `ssp370_Local calibration` = -0.1,
                        `ssp585_Projected` = -0.2,
                        `ssp585_Local calibration` = -0.25),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))) %>% 
  drop_na() %>% 
  ggplot(aes(y = spec_loc)) + 
  geom_point(aes(y = spec_loc + dodge, x  = med_future, col = local_cultivar)) +
  geom_errorbar(aes(y = spec_loc + dodge, xmin = med_future - sd_future, xmax = med_future + sd_future, col = local_cultivar)) +
  facet_grid(species~scenario_year, scales = 'free_y', space = 'free_y') +
  theme_bw(base_size = 15) +
  scale_x_continuous(breaks = c(1, 32,  60, 91, 121, 152), 
                     labels = c('Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun'),
                     minor_breaks = c(1, 32, 60, 91, 121, 152)) +
  scale_y_reverse(breaks = 1:13, 
                  labels = c('Klein-Altendorf', 'Meknes', 'Klein-Altendorf', 'Zaragoza', 'Zaragoza', 'Cieza', 'Klein-Altendorf', 'Klein-Altendorf', 'Zaragoza', 'Santomera', 'Meknes', 'Sfax', 'Sfax'), 
                  minor_breaks = NULL) +
  theme(legend.position = 'bottom') +
  ylab('') +
  xlab('Median Predicted Bloom Date')



fail_plot_df2 <- fail_plot %>% 
  rename(fail_rate = 'median_fail',
         fail_text = 'median_fail_text') %>% 
  select(species, location, ssp, scenario_year, local_cultivar, species_location, spec_loc, fail_rate, fail_text)

x_cutoff <- 170
dodge_x_beginning <- 10
point_size <- 4.5
extra_dodge_x <- 10
gap_between_ssp <- 0.1
gap_between_treatments <- 0.5 * gap_between_ssp


pos_ssp1_local <- (gap_between_ssp * -2) - gap_between_treatments
pos_ssp1_foreign <- gap_between_ssp * -2
pos_ssp2_local <- -gap_between_ssp
pos_ssp2_foreign <- -gap_between_treatments
pos_ssp3_local <- gap_between_treatments 
pos_ssp3_foreign  <- gap_between_ssp
pos_ssp5_local <- (gap_between_ssp * 2)
pos_ssp5_foreign <- (gap_between_ssp * 2) + gap_between_treatments

pos_2015_local <- gap_between_treatments * -1
pos_2015_foreign <- gap_between_treatments


fail_combined <- fail_2015 %>% 
  mutate(ssp = '2015',
         scenario_year = 2015) %>%
  drop_na() %>% 
  rbind(fail_plot_df2) %>% 
  mutate( species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                    "Apple_Meknes",
                                                                                    "Pear_Klein-Altendorf",
                                                                                    "Pear_Zaragoza",
                                                                                    "Apricot_Zaragoza",
                                                                                    "Apricot_Cieza",
                                                                                    "Europ. Plum_Klein-Altendorf",
                                                                                    "Sweet Cherry_Klein-Altendorf",
                                                                                    "Sweet Cherry_Zaragoza",
                                                                                    "Almond_Santomera",
                                                                                    "Almond_Meknes",
                                                                                    "Almond_Sfax",
                                                                                    "Pistachio_Sfax")),
          ssp_local = paste(ssp, local_cultivar, sep = '_'), 
                  dodge = recode(ssp, 
                                 `2015` = mean(c(pos_2015_foreign, pos_2015_local)),
                                 `ssp126` =mean(c(pos_ssp1_foreign, pos_ssp1_local)),
                                 `ssp245` = mean(c(pos_ssp2_foreign, pos_ssp2_local)),
                                 `ssp370` = mean(c(pos_ssp3_local, pos_ssp3_foreign)),
                                 `ssp585` = mean(c(pos_ssp5_foreign, pos_ssp5_local))),
                  dodge_x = recode(local_cultivar, 
                                       `Local calibration` = 0,
                                       `Projected` = extra_dodge_x),
                  species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                         labels = c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
                  x = x_cutoff + dodge_x_beginning) %>% 
  drop_na()




med_plot_df <- median_2015_locfor %>% 
  mutate(ssp = '2015', 
         scenario_year = 2015) %>% 
  rename(`med_future` = 'med_current',
         `sd_future` = 'sd_current') %>% 
  rbind(median_future_local_vs_foreign_allgcm) %>% 
  mutate(local_cultivar = factor(local_cultivar, levels = c(TRUE, FALSE), labels = c('Local calibration', 'Projected')),
         ssp_local = paste(ssp, local_cultivar, sep = '_'), 
         dodge = recode(ssp_local, 
                        `2015_Projected` = pos_2015_foreign,
                        `2015_Local calibration` = pos_2015_local,
                        `ssp126_Projected` = pos_ssp1_foreign,
                        `ssp126_Local calibration` =pos_ssp1_local,
                        `ssp245_Projected` = pos_ssp2_foreign,
                        `ssp245_Local calibration` = pos_ssp2_local,
                        `ssp370_Projected` = pos_ssp3_foreign,
                        `ssp370_Local calibration` = pos_ssp3_local,
                        `ssp585_Projected` = pos_ssp5_foreign,
                        `ssp585_Local calibration` = pos_ssp5_local),
         species_location = factor(paste(species, location, sep = '_'), levels = c("Apple_Klein-Altendorf",
                                                                                   "Apple_Meknes",
                                                                                   "Pear_Klein-Altendorf",
                                                                                   "Pear_Zaragoza",
                                                                                   "Apricot_Zaragoza",
                                                                                   "Apricot_Cieza",
                                                                                   "European Plum_Klein-Altendorf",
                                                                                   "Sweet Cherry_Klein-Altendorf",
                                                                                   "Sweet Cherry_Zaragoza",
                                                                                   "Almond_Santomera",
                                                                                   "Almond_Meknes",
                                                                                   "Almond_Sfax",
                                                                                   "Pistachio_Sfax")),
         spec_loc = as.numeric(species_location),
         species = factor(species, levels =  c('Apple', 'Pear', 'Apricot', 'European Plum', 'Sweet Cherry', 'Almond', 'Pistachio'),
                          labels = c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio'))) %>% 
  drop_na()


df_text_no_pred <- data.frame(y =  rep(c(0.5,2.5, 4.5, 6.5, 7.5, 9.45, 12.5), 2),
                              x = x_cutoff + 2,
                              species = factor(rep(c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio'), 2),
                                               levels =  c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
                              local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 7), levels = c('Local calibration', 'Projected')))

df_padding_box <- data.frame(y = rep(c(6.45, 12.45, 7.55, 13.55), 2),
                             x = x_cutoff +1,
                             species = factor(rep(c('Europ. Plum', 'Pistachio'), 4),
                                              levels =  c('Apple', 'Pear', 'Apricot', 'Europ. Plum', 'Sweet Cherry', 'Almond', 'Pistachio')),
                             local_cultivar = factor(rep(c('Local calibration', 'Projected'), each = 4), levels = c('Local calibration', 'Projected')))

med_plot_df %>% 
  filter(local_cultivar == 'Local calibration') %>% 
  ggplot(aes(y = spec_loc)) + 
  geom_rect(xmin = x_cutoff, xmax = Inf,ymax = Inf, ymin = -Inf, fill = 'grey85',  size = 2) +
  geom_point(data = df_padding_box, aes(y = y, x = x), col = 'grey85', size = 0.1) +
  geom_text(data = df_text_no_pred, aes(y = y, x = x,
                                        label = 'Fail'),hjust = 0) +
  geom_point(aes(y = spec_loc + dodge, x  = med_future), col =  "#56B4E9") +
  geom_errorbar(aes(y = spec_loc + dodge, xmin = med_future - sd_future, xmax = med_future + sd_future), col =  "#56B4E9") +
  geom_point(data = med_plot_df[med_plot_df$local_cultivar == 'Projected',], aes(y = spec_loc + dodge, x  = med_future), col = "#E69F00") +
  geom_errorbar(data = med_plot_df[med_plot_df$local_cultivar == 'Projected',],
                aes(y = spec_loc + dodge, xmin = med_future - sd_future, xmax = med_future + sd_future), col = "#E69F00") +
  geom_point(data = fail_combined, aes(x=x + dodge_x, y= spec_loc + dodge, fill=fail_rate), col = 'black', shape=22, size=point_size) +
  scale_fill_gradient(low = "white", high = "firebrick2", limits = c(0, 100),name = 'Failure Rate') +
  facet_grid(species~scenario_year, scales = 'free_y', space = 'free_y') +
  theme_bw(base_size = 15) +
  scale_x_continuous(breaks = c(1, 32,  60, 91, 121, 152), 
                     labels = c('Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun'),
                     minor_breaks = c(1, 32, 60, 91, 121, 152)) +
  scale_y_reverse(breaks = 1:13, 
                  labels = c('Klein-Altendorf', 'Meknes', 'Klein-Altendorf', 'Zaragoza', 'Zaragoza', 'Cieza', 'Klein-Altendorf', 'Klein-Altendorf', 'Zaragoza', 'Santomera', 'Meknes', 'Sfax', 'Sfax'), 
                  minor_breaks = NULL) +
  theme(legend.position = 'bottom') +
  ylab('') +
  xlab('Median Predicted Bloom Date')
ggsave('figures/paper/local_foreign_median_sd_allyears.jpeg', height = 30, width = 25, units = 'cm', device = 'jpeg')



#make a table with median shifts, median failure rates and median bloom dates
#have sd, min, max aswell



performance %>% 
  group_by(species) %>% 
  summarize(med_rmse = median(rmse))
